Revert "Revert "Merge branch 'plugins'""
This reverts commit 01e6f017bef40927686d78876c1635d60c48e759.
