+++ package.json
@@ -8,7 +8,7 @@
   },
   "scripts": {
     "prepublish": "npm run build -s",
-    "build": "babel src --out-dir lib --sourceRoot=./ --copy-files --presets=es2015,es2016,es2017,stage-3",
+    "build": "babel src --out-dir lib --sourceRoot=./ --copy-files --source-maps --presets=es2015,es2016,es2017,stage-3",
     "build:watch": "npm run build -s -- -w",
     "test": "bash ./tests/run.sh",
     "test:custom-server": "nofat test",
@@ -63,6 +63,7 @@
     "silent-npm-registry-client": "2.x.x",
     "ssh2": "0.4.x",
     "underscore": "1.x.x",
-    "uuid": "2.x.x"
+    "uuid": "2.x.x",
+    "yargs": "^7.0.2"
   }
 }
