+++ src/index.js
@@ -1,157 +1,119 @@
-import MupAPI from './mup-api';
 import checkUpdates from './updates';
-import modules from './modules/';
+import modules, { loadPlugins } from './modules/';
 import pkg from '../package.json';
-import program from 'commander';
+// import program from 'commander';
+import yargs from 'yargs';
+import chalk from 'chalk';
+import MupAPI from './mup-api';
 
-let settingsPath;
-let configPath;
-const args = process.argv.slice(2);
-
-program
-  .arguments('<command> [subcommand]')
-  .version(pkg.version)
-  .action(argAction)
-  .option('--settings <filePath>', 'Meteor settings file', setSettingsPath)
-  .option('--config <filePath>', 'mup.js config file', setConfigPath)
-  .option(
-    '--verbose',
-    'Print more output while building and running tasks on server'
-  )
-  .on('--help', function() {
-    console.log('   Commands:');
-
-    function listModuleCommands(commands) {
-      Object.keys(commands).forEach(command => {
-        if (command === 'default') {
-          listModuleCommands(commands['default']);
-          return;
-        }
-        console.log(`     ${command}`);
-      });
-    }
-
-    listModuleCommands(modules);
-
-    console.log('');
-    console.log('    For list of subcommands, run ');
-    console.log('      mup <command> help');
-  })
-  .parse(process.argv);
-
-if (program.args.length === 0) {
-  program.help();
-  process.exit(0);
-}
-
-function argAction(arg, subarg) {
-  let moduleArg = arg;
-  let command = subarg;
-
-  if (!command && !modules[moduleArg]) {
-    command = moduleArg;
-    moduleArg = 'default';
-  }
-
-  if (moduleArg === 'default' && command === 'help') {
-    program.help();
-    process.exit();
-  }
-
-  let module;
-
-  if (modules[moduleArg]) {
-    module = modules[moduleArg];
-  } else {
-    console.error(`No such module ${moduleArg}`);
-    program.help();
-    process.exit(1);
-  }
-
-  if (!command) {
-    if (moduleArg === 'default') {
-      program.help();
-    } else {
-      module.help(args);
-    }
-    process.exit(0);
-  }
-
-  if (!module[command]) {
-    console.error('error: unknown command %s', command);
-    if (moduleArg === 'default') {
-      program.help();
-      process.exit(1);
-    }
-
-    module.help(args);
-    process.exit(1);
-  }
-
-  if (program.settings) {
-    let settingsIndex = argIndex(args, '--settings');
-
-    if (args[settingsIndex].indexOf('--settings=') === 0) {
-      args.splice(settingsIndex, 1);
-    } else {
-      args.splice(settingsIndex, 2);
-    }
-  }
-
-  if (program.config) {
-    let configIndex = argIndex(args, '--config');
-
-    if (args[configIndex].indexOf('--config=') === 0) {
-      args.splice(configIndex, 1);
-    } else {
-      args.splice(configIndex, 2);
-    }
-  }
-
-  checkUpdates().then(() => {
-    const base = process.cwd();
-    const api = new MupAPI(
-      base,
-      args,
-      configPath,
-      settingsPath,
-      program.verbose
+function addModuleCommands(builder, module) {
+  Object.keys(module.commands).forEach((commandName) => {
+    let command = module.commands[commandName];
+    command.builder = command.builder || {};
+    builder.command(
+        commandName,
+        command.description,
+        command.builder,
+        commandWrapper(command.handler)
     );
-    let potentialPromise = module[command](api);
-    if (potentialPromise && typeof potentialPromise.then === 'function') {
-      potentialPromise.catch(e => {
-        if (e.nodemiralHistory instanceof Array) {
-          // Error is from nodemiral when running a task list
-          // Nodemiral already displayed the error to the user
-          return;
-        }
-
-        console.log(e);
-      });
-    }
   });
 }
 
-function argIndex(list, string) {
-  for (let i = 0; i < list.length; i++) {
-    if (list[i].indexOf(string) === 0) {
-      return i;
+function filterArgv() {
+  const unwanted = [
+    '_',
+    '$0',
+    'settings',
+    'config'
+  ];
+  let result = [...yargs.argv._];
+
+  Object.keys(yargs.argv).forEach((key) => {
+    if (unwanted.indexOf(key) === -1 && yargs.argv[key] !== false && yargs.argv[key] !== undefined) {
+      result.push(`--${key}`);
+
+      if (typeof yargs.argv[key] !== 'boolean') {
+        result.push(yargs.argv[key]);
+      }
     }
+  });
+
+  return result;
+}
+
+function commandWrapper(handler) {
+  return function() {
+    checkUpdates().then(() => {
+      const api = new MupAPI(
+        process.cwd(),
+        filterArgv(),
+        yargs.argv
+      );
+
+      let potentialPromise = handler(api);
+
+      if (potentialPromise && typeof potentialPromise.then === 'function') {
+        potentialPromise.catch(e => {
+          if (e.nodemiralHistory instanceof Array) {
+            // Error is form nodemiral when running a task list.
+            // Nodemiral already displayed the error
+            return;
+          }
+          console.log(e);
+        });
+      }
+    }).catch(e => {
+      console.error(e);
+    });
+  };
+}
+
+// Load plugins
+let config = new MupAPI(process.cwd(), process.argv, yargs.argv).getConfig();
+if (config.plugins) {
+  loadPlugins(config.plugins);
+}
+
+let program = yargs
+  .usage(`\nUsage: ${chalk.yellow('mup')} <command> [args]`)
+  .version(pkg.version)
+  .alias('version', 'V')
+  .global('version', false)
+  .option('settings', {
+    description: 'Path to Meteor settings file',
+    requiresArg: true,
+    string: true
+  })
+  .option('config', {
+    description: 'Path to mup.js config file',
+    requiresArg: true,
+    string: true
+  })
+  .option('verbose', {
+    description: 'Print output from build and server scripts',
+    boolean: true
+  })
+  .strict(true)
+  .alias('help', 'h')
+  .epilogue('For more information, read the docs at https://github.com/zodern/meteor-up')
+  .help('help');
+
+Object.keys(modules).forEach(moduleName => {
+  if (moduleName !== 'default' && modules[moduleName].commands) {
+    yargs.command(
+        moduleName,
+        modules[moduleName].description,
+        (subYargs) => {
+          addModuleCommands(subYargs, modules[moduleName], moduleName);
+        }
+      );
+  } else if (moduleName === 'default') {
+    addModuleCommands(yargs, modules[moduleName], moduleName);
   }
-  return -1;
-}
+});
 
-function handleErrors(e) {
-  console.log(e.name, e.message);
-  process.exit(1);
-}
+program = program.argv;
 
-function setSettingsPath(settingsPathArg) {
-  settingsPath = settingsPathArg;
+if (program._.length === 0) {
+  yargs.showHelp();
 }
-
-function setConfigPath(configPathArg) {
-  configPath = configPathArg;
-}
-
-process.on('uncaughtException', handleErrors);
